pdm_t_dist : Version stable 2

+++ src/struct/pdm_block_to_part.c
@@ -230,8 +230,8 @@ PDM_block_to_part_exch
   unsigned char *sendBuffer = NULL;
   unsigned char *recvBuffer = NULL;
   
-  int s_sendBuffer = 0;
-  int s_recvBuffer = 0;
+  size_t s_sendBuffer = 0;
+  size_t s_recvBuffer = 0;
   
   int s_comm1 = _btp->s_comm - 1;
